\name{dinvGamma}
\alias{dinvGamma}
\title{Density function for the inverse gamma distribution...}
\usage{dinvGamma(x, a, b, log=FALSE, normalize=TRUE)}
\description{Density function for the inverse gamma distribution}
\value{the density values at \code{x}}
\keyword{distribution}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{x}{vector of values in the positive support}
\item{a}{shape parameter}
\item{b}{rate parameter}
\item{log}{logical; if \code{TRUE}, the log density is returned}
\item{normalize}{normalize the density function? (default)}}
