\name{HypergnPrior-initialize}
\alias{HypergnPrior-initialize}
\alias{initialize,HypergnPrior-method}
\title{Initialization method for the "HypergnPrior" class}
\arguments{
  \item{.Object}{the \code{\linkS4class{HypergnPrior}} we
  want to initialize}

  \item{a}{the hyperparameter value}

  \item{n}{the sample size}

  \item{\dots}{unused}
}
\value{
  the initialized object
}
\description{
  Initialization method for the "HypergnPrior" class
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{internal}
\keyword{methods}

