\name{hypergsplines-package}
\title{Bayesian model selection with penalized splines and hyper-g prior...}
\description{Bayesian model selection with penalized splines and hyper-g prior}
\details{This R-package implements Bayesian model selection with penalized
splines, using a hyper-g prior for the linear effects coefficients and the
regression variance in the normal linear and generalised linear models.}
\alias{hypergsplines}
\docType{package}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\keyword{package}
\keyword{regression}
