\name{is.bool}
\alias{is.bool}
\title{Predicate checking for a boolean option}
\usage{
  is.bool(x)
}
\arguments{
  \item{x}{the object being checked}
}
\value{
  Returns \code{TRUE} if \code{x} is a logical scalar
}
\description{
  Predicate checking for a boolean option
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{internal}

