\name{is.posInt}
\alias{is.posInt}
\title{Predicate checking for a scalar positive integer}
\usage{
  is.posInt(x)
}
\arguments{
  \item{x}{the object being checked}
}
\value{
  Returns \code{TRUE} if \code{x} is a scalar positive
  integer.
}
\description{
  Predicate checking for a scalar positive integer
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{internal}

