\name{is.scalar}
\alias{is.scalar}
\title{Predicate checking for a scalar}
\usage{
  is.scalar(x)
}
\arguments{
  \item{x}{the object being checked}
}
\value{
  Returns \code{TRUE} if \code{x} is a length one vector,
  i.e. a scalar.
}
\description{
  Predicate checking for a scalar
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{internal}

