\name{linApproxDens}
\alias{linApproxDens}
\title{Test the C++ object class LinApproxDens...}
\usage{linApproxDens(args, logDens, grid, nSamples)}
\description{Test the C++ object class LinApproxDens}
\value{list with approximate density values and samples}
\keyword{internal}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{args}{the x values}
\item{logDens}{the unnormalised log density values}
\item{grid}{the grid on which to evaluate the linear spline approximation}
\item{nSamples}{number of samples to produce}}
