\name{qinvGamma}
\alias{qinvGamma}
\title{Quantile function for the inverse gamma distribution...}
\usage{qinvGamma(p, a, b, lower.tail=TRUE, log.p=FALSE)}
\description{Quantile function for the inverse gamma distribution}
\value{the quantiles values}
\keyword{distribution}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{p}{vector of (log) probabilities}
\item{a}{shape parameter}
\item{b}{rate parameter}
\item{lower.tail}{logical; if \code{TRUE} (default), probabilities
\code{p} are P[X <= x], otherwise, P[X > x]}
\item{log.p}{logical; if \code{TRUE}, \code{p} are the log probabilities}}
