\name{rinvGamma}
\alias{rinvGamma}
\title{Generate random numbers from the inverse gamma distribution}
\usage{
  rinvGamma(n, a, b)
}
\arguments{
  \item{n}{number of observations. If \code{length(n) > 1},
  the length is taken to be the number required}

  \item{a}{shape parameter}

  \item{b}{rate parameter}
}
\value{
  the random variates
}
\description{
  Generate random numbers from the inverse gamma
  distribution
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{distribution}

