\docType{package}
\name{hypergsplines-package}
\alias{hypergsplines}
\alias{hypergsplines-package}
\title{Bayesian model selection with penalized splines and hyper-g priors}
\description{
  This R-package implements Bayesian model selection with
  penalized splines, using hyper-g priors for the linear
  effects coefficients and the regression variance in the
  normal linear and generalised linear models.
}
\author{
  Daniel Sabanes Bove
  \email{daniel.sabanesbove@ifspm.uzh.ch}
}
\keyword{package}
\keyword{regression}

