\name{is.bool}
\alias{is.bool}
\title{Predicate checking for a boolean option...}
\usage{is.bool(x)}
\description{Predicate checking for a boolean option}
\value{Returns \code{TRUE} if \code{x} is a logical scalar}
\keyword{internal}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{x}{the object being checked}}
