\name{is.posInt}
\alias{is.posInt}
\title{Predicate checking for a scalar positive integer...}
\usage{is.posInt(x)}
\description{Predicate checking for a scalar positive integer}
\value{Returns \code{TRUE} if \code{x} is a scalar positive integer.}
\keyword{internal}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{x}{the object being checked}}
