\name{is.scalar}
\alias{is.scalar}
\title{Predicate checking for a scalar...}
\usage{is.scalar(x)}
\description{Predicate checking for a scalar}
\value{Returns \code{TRUE} if \code{x} is a length one vector, i.e. a
scalar.}
\keyword{internal}
\author{Daniel Sabanes Bove \email{daniel.sabanesbove@ifspm.uzh.ch}}
\arguments{\item{x}{the object being checked}}
